+++ galgebra/utils.py
@@ -4,8 +4,15 @@ Utility Classes
 
 import sys
 import collections
+import warnings
 
 from io import StringIO  # noqa: F401
+
+# galgebra 0.5.0
+warnings.warn(
+    "galgebra.utils is deprecated and will be removed. "
+    "If you need python 2 compatibility helpers, use a decdicated module like "
+    "`six`.", DeprecationWarning, stacklevel=2)
 
 # From https://github.com/benjaminp/six/blob/master/six.py
 PY2 = sys.version_info[0] == 2